ImageTransformation/requirements.txt
certifi
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
filelock==3.9.0
fonttools==4.39.2
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.1
scipy==1.10.1
six==1.16.0
sympy==1.11.1
torch==2.0.0+cu118
torchvision==0.15.1+cu118
tqdm==4.65.0
typing_extensions==4.4.0
urllib3==1.26.13
wincertstore==0.2
